Changes For Approve User
Changes For Approve single User with test cases
