+++ bnd-workspace/de.fraunhofer.abm.collection.dao/src/de/fraunhofer/abm/collection/dao/UserDao.java
@@ -14,7 +14,7 @@ public interface UserDao {
 
 	public void addUser(String username, String firstname, String lastname, String email, String affiliation, String password, String token);
 
-	public void updateUser(String username, String firstname, String lastname, String email, String affiliation, String saltHashPassword);
+	public void updateUser(String username, String firstname, String lastname, String email, String affiliation, String password, String token);
 
 	public String approveToken(String name, String token);
