+++ bnd-workspace/de.fraunhofer.abm.test.http/src/de/fraunhofer/abm/test/http/UserInfoControllerTest.java
@@ -3,7 +3,8 @@ package de.fraunhofer.abm.test.http;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-	
+
+import org.json.JSONObject;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -28,6 +29,7 @@ import de.fraunhofer.abm.http.client.HttpResponse;
 	        testUpdateUser();
 	        testUpdateUserInfoNull();
 	        testRegisterUserNullInfo();
+	        testApproveUser();
 	    }	   
 	    
 		protected Map<String, String> login() throws IOException {
@@ -84,5 +86,34 @@ import de.fraunhofer.abm.http.client.HttpResponse;
 	        response = HttpUtils.post(uri, headers, payload.getBytes(charset), charset);
 	        Assert.assertEquals(NUM200, response.getResponseCode());
 		}
+		
+		private void testApproveUser() throws IOException {
+			HttpResponse response;
+			Map<String, String> headers = login();
+	        headers.put("Content-Type", "application/json;charset=UTF-8");
+	 		// String payload = "{\"isApprove\":\"true\", \"userList\":[\"testUser1\", \"testUser2\"]}";
+	        String payload = "{\"isApprove\":\"true\", \"userList\":\"testUser1\"}";
+	        headers.put("params", payload);
+	        String uri = baseUri + "/rest/approval";
+	        response = HttpUtils.put(uri, headers, payload.getBytes(charset), charset);
+	        Assert.assertEquals(NUM200, response.getResponseCode());
+	        
+	        uri = baseUri + "/rest/username?username=" + "testUser1";
+	        String result = HttpUtils.get(uri, headers, charset);
+	        Assert.assertNotNull(result);
+            JSONObject obj = new JSONObject(result);
+            Assert.assertEquals(true, obj.get("approved"));
+		}
+		 // TODO : Reject user will be uncommented after delete user code is checked in
+		/* private void testRejectUser() throws IOException {
+			HttpResponse response;
+			Map<String, String> headers = login();
+	        headers.put("Content-Type", "application/json;charset=UTF-8");
+	 		String payload = "{\"isApprove\":\"false\", \"userList\":\"testUser2\"}";
+	        headers.put("params", payload);
+	        String uri = baseUri + "/rest/approval";
+	        response = HttpUtils.put(uri, headers, payload.getBytes(charset), charset);
+	        Assert.assertEquals(NUM200, response.getResponseCode());
+		}*/
 		
 	}
